msm: acpuclock-8x60: Use hotplug notifications for CPU, L2, bus, voltage
Voltage and bandwidth requests from a CPU should be dropped when that
CPU is unplugged.  Subscribe to hotplug notifications to accomplish
this, replacing the calls to acpuclock from pm-8x60 in the hotplug path.

Now, a hotplug-removed CPU's clock, PLL, L2 vote, bus bandwidth vote,
and voltage votes are not removed until after the CPU has been shut
down. These settings are all restored to their previous values just
prior to bringing the CPU back online.

Change-Id: I1906bb6fc7e2083c3bbfedf31751c6862e671eea
CRs-Fixed: 280501
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>

ARM: 6528/1: Use CTR for the I-cache line size on ARMv7

The current implementation of the v7_coherent_*_range function assumes
that the D and I cache lines have the same size, which is incorrect
architecturally. This patch adds the icache_line_size macro which reads
the CTR register. The main loop in v7_coherent_*_range is split in two
independent loops or the D and I caches. This also has the performance
advantage that the DSB is moved outside the main loop.

Change-Id: Ibba459ab2b9f0310cdbe02ce9adaedd396bb2175
CRs-fixed: 294396
Reported-by: Kevin Sapp <ksapp@quicinc.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
[smuckle@codeaurora.org: added change to Sc45nm workaround]
Signed-off-by: Steve Muckle <smuckle@codeaurora.org>
Signed-off-by: Jin Hong <jinh@codeaurora.org>

sched_fair: Remove cpu from nohz idle set if not active

Currently, when determining which cpu should do load balancing in nohz
mode, we still consider cpus that have been hotremoved. This is a
problem because, if we decide to try and move the load balancing to a
hotremoved cpu, we will spin and stop doing load balancing at all. This
will also make us cease doing any idle power collapse until the cpu has
been brought back online.

Fix this by removing a cpu from the nohz idle mask if we find that it
isn't active when we consider it for load balancing.

Change-Id: I93a86fdcab88964a452cb3f38e4815db13d0ee6c
CRs-Fixed: 290042
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>